refactor(InputView): 사용자로부터 올바른 값을 입력받는다.
